Move test data factory.
