+++ registration/src/test/java/org/spine3/samples/lobby/registration/order/Given.java
@@ -46,7 +46,7 @@ import static org.spine3.money.Currency.USD;
 import static org.spine3.money.MoneyUtil.newMoney;
 import static org.spine3.samples.lobby.common.util.IdFactory.newConferenceId;
 import static org.spine3.samples.lobby.common.util.IdFactory.newOrderId;
-import static org.spine3.samples.lobby.registration.testdata.TestDataFactory.newPersonalInfo;
+import static org.spine3.samples.lobby.common.util.testdata.TestDataFactory.newPersonalInfo;
 import static org.spine3.samples.lobby.registration.util.Seats.newSeatQuantity;
 
 /**
