+++ registration/src/test/java/org/spine3/samples/lobby/registration/procman/Given.java
@@ -60,7 +60,7 @@ import static org.spine3.protobuf.Durations.ofMinutes;
 import static org.spine3.protobuf.Messages.fromAny;
 import static org.spine3.samples.lobby.common.util.IdFactory.newConferenceId;
 import static org.spine3.samples.lobby.common.util.IdFactory.newOrderId;
-import static org.spine3.samples.lobby.registration.testdata.TestDataFactory.newBoundedContext;
+import static org.spine3.samples.lobby.common.util.testdata.TestDataFactory.newBoundedContext;
 import static org.spine3.samples.lobby.registration.util.Seats.newSeatQuantity;
 
 /**
